requirements.txt
absl-py==0.11.0
ai2thor==2.1.0
aws-requests-auth==0.4.3
backcall==0.2.0
blis==0.7.3
botocore==1.18.18
cachetools==4.2.0
catalogue==1.0.0
certifi==2019.3.9
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.2.2
cycler==0.10.0
cymem==2.0.4
Cython==0.29.21
dataclasses==0.8
decorator==4.4.2
docopt==0.6.2
fast-downward @ https://github.com/MarcCote/downward/archive/faster_replan.zip
filelock==3.0.12
Flask==1.1.2
future==0.18.2
gdown==3.13.0
google-auth==1.24.0
google-auth-oauthlib==0.4.2
grpcio==1.34.0
gym==0.15.4
h5py==2.10.0
hashids==1.3.1
idna==2.8
importlib-metadata==3.1.0
ipython==7.16.1
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.2
jericho==3.0.4
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
jsonpatch==1.27
jsonpointer==2.0
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mementos==1.3.1
mkl-fft==1.1.0
mkl-random==1.1.1
mkl-service==2.3.0
more-itertools==8.6.0
msgpack==1.0.2
murmurhash==1.0.4
networkx==2.3
nltk==3.5
numpy==1.16.2
oauthlib==3.1.0
olefile==0.46
opencv-python==4.0.0.21
packaging==20.4
pandas==1.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pipreqs==0.4.10
plac==1.1.3
plotly==3.10.0
preshed==3.0.4
progressbar2==3.53.1
prompt-toolkit==3.0.8
protobuf==3.13.0
psutil==5.7.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.2
pycparser==2.20
pyglet==1.3.2
Pygments==2.7.2
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
python-utils==2.5.6
pytz==2020.1
PyYAML==5.3.1
pyzmq==20.0.0
regex==2020.11.13
requests==2.21.0
requests-oauthlib==1.3.0
retrying==1.3.3
revtok==0.0.3
rsa==4.6
sacremoses==0.0.43
scipy==1.5.4
sentencepiece==0.1.91
six==1.15.0
spacy==2.3.4
srsly==1.0.4
TatSu==4.4.0
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorboardX==1.8
termcolor==1.1.0
textworld @ https://github.com/MohitShridhar/TextWorld/archive/handcoded_expert_integration.zip
thinc==7.4.3
tokenizers==0.9.4
torch==1.6.0
torchfile==0.1.0
torchvision==0.7.0
tornado==6.1
tqdm==4.31.1
traitlets==4.3.3
transformers==4.2.1
urllib3==1.24.3
visdom==0.1.8.9
vocab==0.0.5
wasabi==0.8.0
wcwidth==0.2.5
websocket-client==0.57.0
Werkzeug==2.0.1
yarg==0.1.9
zipp==3.4.0
